refactor: add similar interfaces
